backend/requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
pymongo==4.5.0
redis==5.0.1
bcrypt==4.0.1
cryptography==41.0.7
python-dotenv==1.0.0
marshmallow==3.20.1
pytest==7.4.3
pytest-flask==1.3.0
gunicorn==21.2.0
Werkzeug==2.3.7